requirements.txt
numpy
pandas
scikit-learn
mysql-connector-python
python-dotenv
matplotlib
seaborn
-e .